# server/qualiva-openapi.yaml
openapi: 3.0.3
info:
  title: Qualiva - Enterprise Secure Module Management System (ESMMS) API
  description: |
    Comprehensive OpenAPI 3.x specification for the Qualiva ESMMS backend services.
    Generated strictly via comprehensive source code inspection of `server/src` for API security auditing with OWASP ZAP.
    
    Includes role-based access control (RBAC), multi-engine database proxies, granular moderator permissions,
    proctored assessment workflows, and real-time enterprise communication endpoints.
  version: 1.0.0
servers:
  - url: http://host.docker.internal:5050
    description: Local development backend server
  - url: /
    description: Current host relative path
...
paths:
...
  /api/v1/transfers:
    get:
      tags:
        - Transfers & SLA
      summary: List Course Transfer & SLA Requests
      description: Lists transfer and SLA extension requests with status counts.
      operationId: listTransfers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, PENDING, APPROVED, REJECTED]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, COURSE_TRANSFER, SLA_EXTENSION, DEPARTMENT_TRANSFER]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Transfers list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  counts:
                    type: object
                    properties:
                      pending:
                        type: integer
                      approved:
                        type: integer
                      rejected:
                        type: integer
                      total:
                        type: integer
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransferRequest'
    post:
      tags:
        - Transfers & SLA
      summary: Submit Course Transfer or SLA Extension Request
      description: Student submits request to change tracks, request SLA deadline extension, or relocate department.
      operationId: createTransfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                type:
                  type: string
                  enum: [COURSE_TRANSFER, SLA_EXTENSION, DEPARTMENT_TRANSFER]
                  default: COURSE_TRANSFER
                courseId:
                  type: string
                targetCourseId:
                  type: string
                fromDepartment:
                  type: string
                toDepartment:
                  type: string
                reason:
                  type: string
                  example: "Requesting 14-day SLA deadline extension due to medical leave."
                requestedDays:
                  type: integer
                  default: 14
      responses:
        '201':
          description: Transfer request created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  transfer:
                    $ref: '#/components/schemas/TransferRequest'
        '400':
          description: Reason is mandatory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/transfers/{id}/approve:
    patch:
      tags:
        - Transfers & SLA
      summary: Authorize / Approve Transfer Request
      description: |
        Authorized by ADMIN or MODERATOR roles. If the request is an SLA extension,
        automatically increments the student's assignment `dueDate` by `requestedDays`.
      operationId: approveTransfer
      security:
        - bearerAuth: []
      x-required-roles:
        - ADMIN
        - MODERATOR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  example: "Authorized SLA extension for 14 days."
      responses:
        '200':
          description: Request approved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  transfer:
                    $ref: '#/components/schemas/TransferRequest'
        '400':
          description: Request is not in PENDING status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transfer request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/transfers/{id}/reject:
    patch:
      tags:
        - Transfers & SLA
      summary: Decline Transfer or SLA Extension Request
      description: Authorized by ADMIN or MODERATOR roles. Rejects request and records review notes.
      operationId: rejectTransfer
      security:
        - bearerAuth: []
      x-required-roles:
        - ADMIN
        - MODERATOR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "Extension cannot exceed maximum cohort timeline."
      responses:
        '200':
          description: Request rejected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  transfer:
                    $ref: '#/components/schemas/TransferRequest'
        '403':
          description: Forbidden role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transfer request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# =========================================================================
# Reusable Components & Schemas
# =========================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JSON Web Token passed in the `Authorization` header as `Bearer <token>`.
        Payload contains `sub` (User UUID), `email`, `role`, `status`, and `department`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: esmms_refresh_token
      description: |
        HttpOnly cookie containing the cryptographic refresh token used exclusively by `/api/v1/auth/refresh` and `/api/v1/auth/logout`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request payload."
        details:
          type: array
          items:
            type: string
...
    TransferRequest:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
          example: "TR-1082"
        studentId:
          type: string
        studentName:
          type: string
        studentEmail:
          type: string
        department:
          type: string
          nullable: true
        type:
          type: string
          enum: [COURSE_TRANSFER, SLA_EXTENSION, DEPARTMENT_TRANSFER]
        typeLabel:
          type: string
        currentCourse:
          type: string
        targetCourse:
          type: string
          nullable: true
        reason:
          type: string
        requestedDays:
          type: integer
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        reviewedById:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time